server-ai/requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
Flask==2.2.3
h11==0.14.0
idna==3.4
imageio==2.27.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy_loader==0.2
MarkupSafe==2.1.2
networkx==3.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pydantic==1.10.7
pytesseract==0.3.10
python-dotenv==1.0.0
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.10.1
sniffio==1.3.0
starlette==0.26.1
tifffile==2023.3.21
typing_extensions==4.5.0
uvicorn==0.21.1
Werkzeug==2.2.3
